WIP: modify auth_screen to use the AuthService
this is a good stepping point to commit but I still have to do
more work on the other screens for them to use the services.

I have to add more services because the main service 'WebSocketService'
is now not accesible, so I need to refactor the rest of the screens.